requirements.txt
fastapi
uvicorn
pillow
numpy==1.24.4
tensorflow-cpu==2.17.1
shap
matplotlib
python-multipart
<<<<<<< Updated upstream
opencv-python-headless==4.11.0.86
=======
opencv-python-headless==4.11.0.86
>>>>>>> Stashed changes
